Handle settings in own file
Add support for a --config/-c flag.

Handle creating a settings object in its own file. If a config file is
given then we use that to configure the server. Any missing settings are
gotten from the default settings. Further cli flags are ignored when a
config is given.

Add --help flag to show help message and exit

Add --version flag to show server version and exit

Add unit tests for the settings handler.